+++ clickable-tags/manifest.json
@@ -1,0 +1,1 @@
+{"package": "clickable_tags_2.1.20+", "name": "Clickable tags 2.1.20+"}
